+++ R/dohle_siegrist_w.R
@@ -8,7 +8,7 @@
 #'   complex drug names are perceived as more hazardous, which makes someone
 #'   less likely to buy the drug. Researchers used a within-subject design.
 #'
-#'   This data set is in a wide format, see \code{\link{mdt_within_wide}} to
+#'   This data set is in a wide format, see [`mdt_within_wide`] to
 #'   conduct a within-participant mediation analysis with this dataset.
 #'
 #'@format A data frame with 22 rows and 5 variables: 
@@ -22,7 +22,7 @@
 #'
 #'@references Dohle, S., & Siegrist, M. (2014). Fluency of pharmaceutical drug
 #'  names predicts perceived hazardousness, assumed side effects and willingness
-#'  to buy. \emph{Journal of Health Psychology}, \emph{19}(10), 1241-1249. doi:
+#'  to buy. *Journal of Health Psychology*, *19*(10), 1241-1249. doi:
 #'  10.1177/1359105313488974
 #'
 #'@usage data("dohle_siegrist_wide")