+++ R/mdt_simple.R
@@ -1,7 +1,7 @@
 #' @title Joint-significance test for simple mediation
 #'
-#' @description Given a data frame, a predictor (\code{IV}), an outcome
-#'   (\code{DV}), and a mediator (\code{M}), conducts a joint-significant test 
+#' @description Given a data frame, a predictor (`IV`), an outcome
+#'   (`DV`), and a mediator (`M`), conducts a joint-significant test 
 #'   for simple mediation (see Yzerbyt, Muller, Batailler, & Judd, 2018).
 #'
 #' @param data A data frame containing the variables to be used in the model.
@@ -44,9 +44,9 @@
 #'   - \eqn{Y_i = b_{30} + \mathbf{c'_{31}} X_i + \mathbf{b_{32}} M_i}{Yi =
 #'     b_30 + c'_31*Xi + b_32*Mi}
 #'
-#'   with \eqn{Y_i}{Yi}, the outcome value for the \emph{i}th observation,
-#'   \eqn{X_i}{Xi}, the predictor value for the \emph{i}th observation, and
-#'   \eqn{M_i}{Mi}, the mediator value for the \emph{i}th observation (Cohen &
+#'   with \eqn{Y_i}{Yi}, the outcome value for the *i*th observation,
+#'   \eqn{X_i}{Xi}, the predictor value for the *i*th observation, and
+#'   \eqn{M_i}{Mi}, the mediator value for the *i*th observation (Cohen &
 #'   Cohen, 1983; Yzerbyt, Muller, Batailler, & Judd, 2018).
 #'   
 #'   Coefficients associated with \eqn{a}, \eqn{b}, \eqn{c}, and \eqn{c'} paths
@@ -55,20 +55,20 @@
 #'   
 #' @section Variable coding: Because joint-significance tests uses linear models
 #'   behind the scenes, variables involved in the model have to be numeric.
-#'   \code{mdt_simple} will give an error if non-numeric variables are
+#'   `mdt_simple` will give an error if non-numeric variables are
 #'   specified in the model.
 #'
 #'   To convert a dichotomous categorical variable to a numeric one, please
-#'   refer to the \code{\link{build_contrast}} function.
+#'   refer to the [`build_contrast`] function.
 #'
-#' @references Cohen, J., & Cohen, P. (1983). \emph{Applied multiple
-#'   regression/correlation analysis for the behavioral sciences} (2nd ed).
+#' @references Cohen, J., & Cohen, P. (1983). *Applied multiple
+#'   regression/correlation analysis for the behavioral sciences* (2nd ed).
 #'   Hillsdale, N.J: L. Erlbaum Associates.
 #'
 #'   Yzerbyt, V., Muller, D., Batailler, C., & Judd, C. M. (2018). New
 #'   recommendations for testing indirect effects in mediational models: The
-#'   need to report and test component paths. \emph{Journal of Personality and
-#'   Social Psychology}, \emph{115}(6), 929–943. doi: 10.1037/pspa0000132
+#'   need to report and test component paths. *Journal of Personality and
+#'   Social Psychology*, *115*(6), 929–943. doi: 10.1037/pspa0000132
 #' 
 #' @examples
 #' ## fit a simple mediation model
@@ -105,34 +105,34 @@ mdt_simple.data.frame <- function(data, IV, DV, M) {
 
   # type check ----------------------------------------------------------------
   if(!is.numeric(IV_data))
-    stop(glue::glue("Warning:
-                    IV ({IV_name}) must be numeric (see build_contrast() to
-                    convert a character vector to a contrast code)."))
+    stop(glue("Warning:
+               IV ({IV_name}) must be numeric (see build_contrast() to
+               convert a character vector to a contrast code)."))
 
   if(!is.numeric(M_data))
-    stop(glue::glue("Warning:
-                    Mediator ({M_name}) must be numeric."))
+    stop(glue("Warning:
+               Mediator ({M_name}) must be numeric."))
 
   if(!is.numeric(DV_data))
-    stop(glue::glue("Warning:
-                    DV ({DV_name}) must be numeric."))
+    stop(glue("Warning:
+               DV ({DV_name}) must be numeric."))
 
   # building models -----------------------------------------------------------
   model1 <-
-    stats::as.formula(glue::glue("{DV} ~ {IV}",
-                                 IV = IV_name,
-                                 DV = DV_name))
+    stats::as.formula(glue("{DV} ~ {IV}",
+                           IV = IV_name,
+                           DV = DV_name))
 
   model2 <-
-    stats::as.formula(glue::glue("{M} ~ {IV}",
-                                 IV = IV_name,
-                                 M  = M_name))
+    stats::as.formula(glue("{M} ~ {IV}",
+                           IV = IV_name,
+                           M  = M_name))
 
   model3 <-
-    stats::as.formula(glue::glue("{DV} ~ {IV} + {M}",
-                                 DV = DV_name,
-                                 IV = IV_name,
-                                 M  = M_name))
+    stats::as.formula(glue("{DV} ~ {IV} + {M}",
+                           DV = DV_name,
+                           IV = IV_name,
+                           M  = M_name))
 
   # models fitting and cleaning -----------------------------------------------
   js_models <-