+++ R/mdt_within_index.R
@@ -2,10 +2,10 @@
 #'
 #' @description Adds the confidence interval for the index of
 #'   within-participant mediation to a  model fitted with
-#'   \code{\link{mdt_within}} or \code{\link{mdt_within_wide}}.
-#'   
+#'   [`mdt_within`] or [`mdt_within_wide`].
+#'  
 #' @param mediation_model A mediation model of class
-#'   \code{"within_participant_mediation"}.
+#'   `"within_participant_mediation"`.
 #' @param times Number of simulations to use to compute the Monte Carlo indirect
 #'   effect confidence interval.
 #' @param level Alpha threshold to use for the confidence interval.
@@ -18,7 +18,7 @@
 #'
 #' @references MacKinnon, D. P., Lockwood, C. M., & Williams, J. (2004).
 #'   Confidence Limits for the Indirect Effect: Distribution of the Product and
-#'   Resampling Methods. \emph{Multivariate Behavioral Research}, \emph{39}(1),
+#'   Resampling Methods. *Multivariate Behavioral Research*, *39*(1),
 #'   99-128. doi: 10.1207/s15327906mbr3901_4
 #'
 #' @examples