+++ R/utils-glue.R
@@ -1,0 +1,17 @@
+#' Format and interpolate a string
+#' 
+#' @description 
+#' 
+#' * [`glue::glue`] formats and interpolates a string. Expressions enclosed by
+#'   braces will be evaluated as R code. Long strings are broken by line and
+#'   concatenated together. Leading whitespace and blank lines from the first
+#'   and ast lines are automatically trimmed.
+#' 
+#' @name glue
+
+#' @keywords internal
+#' @noRd
+#'
+#' @importFrom glue glue
+#' @aliases glue
+NULL